feature table changes
